Fix syncing errors due to new products since Nov. 2019
Problems with sync were caused primarily by new categories of products being created: The Mystery Booster Convention Edition "Test Print" cards, the Secret Lair Drop Series, and the Godzilla series of promos for Ikoria.

Godzilla was solved by using a mapper to connect the Godzilla name to its base card name for scryfall. The rest was mostly one-off exceptions added to constants similar to booster boxes.

The last class of error was the most annoying: Typos by TCGPlayer. See the STUPID_TCGPLAYER_TYPO_MAPPER constant to see what that idiocy is about. This was the best solution as it will not break this databse if TCGPlayer fixes these items in the future.